Changed explosives to use LONG type for energy

+++ src/main/java/icbm/api/explosion/IExplosion.java
@@ -13,5 +13,5 @@ public interface IExplosion
 
     /** @return The energy emitted by this explosive. In Joules and approximately based off of a real
      * life equivalent. */
-    public float getEnergy();
+    public long getEnergy();
 }
